requirements.txt
# Deep Learning
torch
torchmetrics
pytorch-lightning

# Datasets & models
datasets
transformers

# Inference
onnxruntime-gpu

# Others
rich
wandb
textual
pre-commit
